Add auth by token
