+++ app/src/main/java/com/softdesign/devintensive/data/network/res/UserRes.java
@@ -1,0 +1,20 @@
+package com.softdesign.devintensive.data.network.res;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class UserRes {
+
+    @SerializedName("success")
+    @Expose
+    private boolean success;
+
+    public User getData() {
+        return data;
+    }
+
+    @SerializedName("data")
+    @Expose
+    private User data;
+
+}
